[Feat] #28 Extension Ptotocol 함수 분리
