+++ Whistle/ViewModels/Protocols/SettingProtocol.swift
@@ -1,0 +1,17 @@
+//
+//  SettingProtocol.swift
+//  Whistle
+//
+//  Created by ChoiYujin on 9/8/23.
+//
+
+import Foundation
+
+protocol SettingProtocol {
+
+  func requestNotiSetting() async
+  func updateSettingWhistle(newSetting: Bool) async
+  func updateSettingFollow(newSetting: Bool) async
+  func updateSettingInfo(newSetting: Bool) async
+  func updateSettingAd(newSetting: Bool) async
+}
