+++ Whistle/ViewModels/Protocols/UploadProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  UploadProtocol.swift
+//  Whistle
+//
+//  Created by ChoiYujin on 9/8/23.
+//
+
+import Foundation
+import UIKit
+
+protocol UploadProtocol {
+  func uploadPhoto(image: UIImage, completion: @escaping (String) -> Void) async
+}
